Make pages pass observables, render header/footer
Having a `<PageWrapper />` obscures what the page renders. Removing it
means pages as a whole are more verbose, but are able to decide whether
they want to render a header or footer.

Pages also pass observables down to section, rather than relying on
context from `@inject()` to render. This means testing is easier as
well, since a section can expect the interface to a store via props, and
not have to concern itself with context.