Refine sections, service behaviour handle errors
Sections currently encompass form logic and interact with observable
stores. It would be useful to break this down into a container and a
display component though to test in isolation.

All methods for the IdentityService now include error handling, which
sets a rejection flag. They then throw the error again, as it is better
for components like a form to handle errors on their rather than through
an observable. I could probably be stricter about this.

Read of the day - https://overreacted.io/writing-resilient-components/