Lift logic up to a page level
I've tried to make section rely on props for more complex behaviour. For
example, authenticating will affect the entire application, not just the
`<SignInForm/>`, so making a service to handle this logic means that the
form component itself only needs to worry about taking a user's input
and passing it to a `signIn()` prop.

This is much easier to test, since I can mock any callback props needed.