Resolve CSRF issues
Requests made through the webpack server will now proxy to the backend,
and the CSRF token Django expects is now passed as a request header.

The CSRF middleware could probably be removed to avoid these problems,
but I liked finding a solution, and I'd prefer to leave Django security
defaults in place.