Decompose views into controllers and services
Routing logic is handled by a set of controllers, which can more
intelligently process requests and hand off the logic to services.

This also changes some frontend routes, since controllers can have
different behaviour dependening on the method of a request. For example,
calling /api/teams with a GET or POST will either return the session
user's team or create a new one for them.