Rewrote MobX stores and did general refactoring
I've rewritten the stores to keep focus on individual domain objects.
In the case of the TeamStore, some logic around retrieving teams is
obscured to allow a hidden cache. At the moment, this cache can become
stale. So it might be worth finding a way to refresh it occasionally.

There's also a lot of minor changes as well here.