Add Docker Compose for backend development
This allows development in an environment that resembles production.

In the case of users developing only on the frontend/backend, there are
two scripts to prepare the frontend and backend for development, so a
contributor does not need to follow a long set of steps to get the
complete application running.