Tidy README, rewrite networking logic
There isn't really a need for typing a separate INetworkInterface, so I
cut it out.

I also added a little useNetwork() function for the sake of ergonmics.
It means that I could strip out the NetworkContext logic as well (say we
wanted to use jest.mock() instead), but that's not a concern for now.

Needed to rejig a little of requireAuthentication() to get TS working.
This means most pages were also touched, escpecially where connect() was
used.