Fix CSRF preventing login, extend E2E test
Since Django enforces CSRF, it's important that the user has a
token when making API requests. I've made sure this is obtained by
adding the `ensure_csrf_cookie` decorator to `GET /api/whoami`, which is
called by the frontend on page load. This should ensure they have the
token when signing in.

Before this, creating an account from a fresh browser would lead to a
403 due to CSRF issues. I updated the E2E to check this. It now gives
more output to help with debugging as well (document, screenshot).

Interestingly, XPath struggled to find text when it was spread across
multiple nodes. I fixed it by combining the nodes in a formatted string.