Modify the typing of meeting venues
Makes it a non-nullable string, with a default of an empty string.

I wasn't consistently checking behaviour around optional field, so I'm
the null-ability. The frontend was not excluding the field from requests
if it was empty, so this is consolidating that behaviour.

In a future (breaking) release, we might make it null-able again, but
that is not an immediate concern.